Disable libxml2
To improve security. In their own words:

"This is open-source software written by hobbyists, maintained by a
single volunteer, badly tested, written in a memory-unsafe language
and full of security bugs. It is foolish to use this software to
process untrusted data. As such, we treat security issues like any
other bug."